renamed SolutionPool objects
