+++ riskslim/lattice_cpa.py
@@ -9,7 +9,7 @@ from .heuristics import discrete_descent, sequential_rounding
 from .initialization import initialize_lattice_cpa
 from .mip import add_mip_starts, convert_to_risk_slim_cplex_solution, create_risk_slim, set_cplex_mip_parameters
 from .setup_functions import get_loss_bounds, setup_loss_functions, setup_objective_functions, setup_penalty_parameters
-from .solution_classes import SolutionPool, SolutionQueue
+from .solution_pool import SolutionPool, FastSolutionPool
 
 DEFAULT_BOUNDS = {
     'objval_min': 0.0,
@@ -315,8 +315,8 @@ def finish_lattice_cpa(data, constraints, mip_objects, settings = DEFAULT_LCPA_S
         'n_bound_updates_objval_max': 0,
         }
 
-    lcpa_cut_queue = SolutionQueue(P)
-    lcpa_polish_queue = SolutionQueue(P)
+    lcpa_cut_queue = FastSolutionPool(P)
+    lcpa_polish_queue = FastSolutionPool(P)
 
     heuristic_flag = lcpa_settings['round_flag'] or lcpa_settings['polish_flag']
 
@@ -477,18 +477,18 @@ class LossCallback(LazyConstraintCallback):
         # setup pointer to cut_queue to receive cuts from PolishAndRoundCallback
         if self.settings['add_cuts_at_heuristic_solutions']:
             if cut_queue is None:
-                self.cut_queue = SolutionQueue(len(self.rho_idx))
+                self.cut_queue = FastSolutionPool(len(self.rho_idx))
             else:
-                assert isinstance(cut_queue, SolutionQueue)
+                assert isinstance(cut_queue, FastSolutionPool)
                 self.cut_queue = cut_queue
 
 
         # setup pointer to polish_queue to send integer solutions to PolishAndRoundCallback
         if self.settings['polish_flag']:
             if polish_queue is None:
-                self.polish_queue = SolutionQueue(len(self.rho_idx))
+                self.polish_queue = FastSolutionPool(len(self.rho_idx))
             else:
-                assert isinstance(polish_queue, SolutionQueue)
+                assert isinstance(polish_queue, FastSolutionPool)
                 self.polish_queue = polish_queue
 
         # setup indices for update bounds
@@ -651,8 +651,8 @@ class PolishAndRoundCallback(HeuristicCallback):
         assert isinstance(indices, dict)
         assert isinstance(control, dict)
         assert isinstance(settings, dict)
-        assert isinstance(cut_queue, SolutionQueue)
-        assert isinstance(polish_queue, SolutionQueue)
+        assert isinstance(cut_queue, FastSolutionPool)
+        assert isinstance(polish_queue, FastSolutionPool)
         assert callable(get_objval)
         assert callable(get_L0_norm)
         assert callable(is_feasible)
@@ -810,7 +810,7 @@ class PolishAndRoundCallback(HeuristicCallback):
             self.polish_queue.filter_sort_unique(max_objval = polishing_cutoff)
 
             if len(self.polish_queue) > 0:
-                polished_queue = SolutionQueue(self.polish_queue.P)
+                polished_queue = FastSolutionPool(self.polish_queue.P)
                 polish_time = 0
                 n_polished = 0
                 for objval, solution in zip(self.polish_queue.objvals, self.polish_queue.solutions):
